+++ apps/scoutgame/components/draft/components/BidButton.tsx
@@ -1,21 +1,13 @@
 'use client';
 
 import { Button, Tooltip } from '@mui/material';
-import { DateTime } from 'luxon';
+import { isDraftEnabled } from '@packages/scoutgame/drafts/checkDraftDates';
 
 import { useGlobalModal } from 'components/common/ModalProvider';
 
-// Draft ends at midnight (end of day) UTC on April 25th, 2024
-const DRAFT_END_DATE = DateTime.fromISO('2025-04-25T23:59:59.999Z', { zone: 'utc' });
-
-export function isDraftEnded(): boolean {
-  const nowUtc = DateTime.utc();
-  return nowUtc > DRAFT_END_DATE;
-}
-
 export function BidButton({ developerPath }: { developerPath: string }) {
   const { openModal } = useGlobalModal();
-  const draftEnded = isDraftEnded();
+  const draftEnabled = isDraftEnabled();
 
   const button = (
     <Button
@@ -29,15 +21,15 @@ export function BidButton({ developerPath }: { developerPath: string }) {
       }}
       onClick={() => openModal('draftDeveloper', { path: developerPath })}
       color='secondary'
-      disabled={draftEnded}
+      disabled={!draftEnabled}
     >
       Bid
     </Button>
   );
 
-  if (draftEnded) {
+  if (!draftEnabled) {
     return (
-      <Tooltip title='Draft has ended' placement='top'>
+      <Tooltip title='Draft has closed or not started yet' placement='top'>
         <div>{button}</div>
       </Tooltip>
     );
