+++ package.json
@@ -46,6 +46,7 @@
     "draft:start:test:ci": "npm run start:test:ci -w=apps/draft",
     "draft": "REACT_APP_IS_DRAFT_SEASON=true npx react-env --path .env.local --dest ./apps/scoutgame/public -- npm run start --workspace=apps/scoutgame",
     "draft:onchain": "REACT_APP_IS_DRAFT_SEASON=true npx react-env --path .env.onchain --dest ./apps/scoutgame/public -- npm run start --workspace=apps/scoutgame",
+    "draft:staging": "REACT_APP_IS_DRAFT_SEASON=true npx react-env --path .env.staging.local --dest ./apps/scoutgame/public -- npm run start --workspace=apps/scoutgame",
     "cron:test": "dotenv -e .env.test.local -- npm test -w apps/cron",
     "cron:staging": "npm run start:prod:node18 -w apps/cron",
     "cron:prod": "npm run start:prod:node18 -w apps/cron",
